% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndStats.R
\name{calcN50}
\alias{calcN50}
\title{Calculate N50 for cumulative reads}
\usage{
calcN50(indf, nfactor = 0.5)
}
\arguments{
\item{indf}{DataFrame of read metadata}

\item{nfactor}{cumulative percentile, default=0.5}
}
\value{
a tibble of results
}
\description{
Uses metadata DataFrame and returns the N50 or any other cumulative percentile
}
