% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndStats.R
\name{createTextSummary}
\alias{createTextSummary}
\title{Calculate read length distribution and coverage then save to .csv}
\usage{
createTextSummary(indf, outputname = NA, append = FALSE)
}
\arguments{
\item{indf}{DataFrame of read metadata}

\item{outputname}{filename (.csv) for export, including path is not saved to
working directory}

\item{append}{TRUE or FALSE to append summary statistics to an existing file}
}
\value{
a vector containing results
}
\description{
Uses metadata DataFrame to calculate read distribution statistics and saves
to file. Optionally append results to existing file where row
name = sample name.
}
