% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndStats.R
\name{loadTable}
\alias{loadTable}
\title{Read in metadata.txt}
\usage{
loadTable(infile, qualityCutoff = 0)
}
\arguments{
\item{infile}{Path to the input file}

\item{qualityCutoff}{reads with quality scores >= to this cutoff are kept}
}
\value{
a DataFrame of the infile
}
\description{
Reads in metadata.txt produced by NanoR and returns a Data.Frame object
Optionally, reads below a quality threshold can be dropped. By default all
reads are returned
}
