% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndStats.R
\name{plotCumulativeBases}
\alias{plotCumulativeBases}
\title{Plot bar chart of cumulative read count}
\usage{
plotCumulativeBases(
  df,
  bin.num,
  plotQual = TRUE,
  qual.bins = c(10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60),
  binWidth = 6000,
  barColor = "#8ec462",
  barColorRamp = c("#8ec462", "firebrick1"),
  colorMapVar = NA,
  dodgeAdjust = 10000
)
}
\arguments{
\item{df}{DataFrame of read metadata}

\item{bin.num}{Number of Bins}

\item{plotQual}{TRUE or FALSE, default FALSE, plot quality distribution as fill}

\item{qual.bins}{a vector of quality cutoffs for use with plotQual}

\item{binWidth}{the width of each bar in b}

\item{barColor}{the name or hex string of a color if not plotting quality}

\item{barColorRamp}{a vector of rcolors to be used when mapping two datasets}

\item{colorMapVar}{name of a column in df to map fill to}

\item{dodgeAdjust}{a number to adjust the spacing of color mapped bars. Default is 10kb.}
}
\value{
a ggplot2 plot
}
\description{
Uses metadata DataFrame to plot the distribution of reads in ggplot. Can plot
read quality in stacked bars, manipulate the width of bins.
}
