% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndStats.R
\name{plotReadCountDistributions}
\alias{plotReadCountDistributions}
\title{Plot bar chart of read counts or cumulative read KB in bins}
\usage{
plotReadCountDistributions(
  indf,
  outname,
  binNum = 50,
  counts = FALSE,
  skinnyBin = FALSE,
  outwidth = 8,
  outheight = 7,
  outunits = "in",
  plotQual = FALSE,
  qual.bins = c(10, 15, 20, 25, 30, 35, 40, 45, 50, 55, 60),
  barColor = "#8ec462",
  indf2 = NA,
  barColorRamp = c("#8ec462", "firebrick1"),
  dodgeAdjust = 10000,
  sampleNames = c("Sample 1", "Sample 2"),
  binWidth = NA
)
}
\arguments{
\item{indf}{DataFrame of read metadata}

\item{outname}{Filename for plot to save. Can be .pdf or .png. Set to NA to
skip plot saving.}

\item{binNum}{Number of Bins}

\item{counts}{TRUE or FALSE, default FALSE, plot counts, not cumulative KB}

\item{skinnyBin}{TRUE or FALSE, default FALSE, make bars skinny for overlay}

\item{outwidth}{width dimension to save plot}

\item{outheight}{height dimension to save plot}

\item{outunits}{unit of width and height, default 'in'}

\item{plotQual}{TRUE or FALSE, default FALSE, plot quality distribution as fill}

\item{qual.bins}{a vector of quality cutoffs for use with plotQual}

\item{barColor}{an r color or color hex code}

\item{indf2}{DataFrame of read metadata for a second sample}

\item{dodgeAdjust}{a number to adjust the spacing of color mapped bars. Default is 10kb.}

\item{sampleNames}{a vector of samplenames for plotting two datasets}

\item{binWidth}{a number, representing width of bins. overrides skinnybin.}

\item{colorMapVar}{name of a column in df to map fill to}
}
\value{
a ggplot2 plot
}
\description{
Uses metadata DataFrame to plot the distribution of reads in ggplot. Can plot
read quality in stacked bars, manipulate the width of bins.
}
